REFACT: Split out of `Examples` project `Cloud` and `OnPremise` parts.
