+++ Examples/Cloud/GetAllProperties-Console/GetAllProperties.csproj
@@ -78,6 +78,7 @@
   </ItemGroup>
 
   <ItemGroup>
+    <PackageReference Include="FiftyOne.IpIntelligence" Version="4.5.0-alpha.120" />
     <PackageReference Include="System.Text.Json" Version="8.0.6" />
   </ItemGroup>
