+++ Examples/Cloud/GettingStarted-Web/Controllers/HomeController.cs
@@ -29,8 +29,6 @@ namespace FiftyOne.IpIntelligence.Examples.Cloud.GettingStartedWeb.Controllers
 {
     public class HomeController : Controller
     {
-        private static bool _checkedDataFile = false;
-
         private IFlowDataProvider _provider;
         private ILogger<HomeController> _logger;
 
@@ -44,12 +42,6 @@ namespace FiftyOne.IpIntelligence.Examples.Cloud.GettingStartedWeb.Controllers
 
         public IActionResult Index()
         {
-            // Log warnings if the data file is too old or the 'Lite' file is being used.
-            if(_checkedDataFile == false)
-            {
-                ExampleUtils.CheckDataFile(_provider.GetFlowData().Pipeline, _logger);
-                _checkedDataFile = true;
-            }
             // Use the provider to get the flow data. This contains the results of device
             // detection that has been performed by the pipeline.
             return View(new IndexModel(_provider.GetFlowData(), Response.Headers));
