+++ Examples/FiftyOne.IpIntelligence.Examples.Cloud/PipelineOptionsExtensions.cs
@@ -1,0 +1,79 @@
+/* *********************************************************************
+ * This Original Work is copyright of 51 Degrees Mobile Experts Limited.
+ * Copyright 2025 51 Degrees Mobile Experts Limited, Davidson House,
+ * Forbury Square, Reading, Berkshire, United Kingdom RG1 3EU.
+ *
+ * This Original Work is licensed under the European Union Public Licence
+ * (EUPL) v.1.2 and is subject to its terms as set out below.
+ *
+ * If a copy of the EUPL was not distributed with this file, You can obtain
+ * one at https://opensource.org/licenses/EUPL-1.2.
+ *
+ * The 'Compatible Licences' set out in the Appendix to the EUPL (as may be
+ * amended by the European Commission) shall be deemed incompatible for
+ * the purposes of the Work and the provisions of the compatibility
+ * clause in Article 5 of the EUPL shall not apply.
+ *
+ * If using the Work as, or as part of, a network application, by
+ * including the attribution notice(s) required under Article 5 of the EUPL
+ * in the end user terms of the application under an appropriate heading,
+ * such notice(s) shall fulfill the requirements of that article.
+ * ********************************************************************* */
+
+using FiftyOne.IpIntelligence.Cloud.FlowElements;
+using FiftyOne.Pipeline.CloudRequestEngine.FlowElements;
+using FiftyOne.Pipeline.Core.Configuration;
+
+namespace FiftyOne.IpIntelligence.Examples.Cloud
+{
+    public static class PipelineOptionsExtensions
+    {
+        /// <summary>
+        /// The name of the setting that is used to specify the resource key in configuration
+        /// files.
+        /// </summary>
+        private const string RESOURCE_KEY_SETTING_NAME = "ResourceKey";
+        /// <summary>
+        /// The name of the setting that is used to specify the cloud end point in configuration
+        /// files.
+        /// </summary>
+        private const string CLOUD_END_POINT_SETTING_NAME = "EndPoint";
+
+        /// <summary>
+        /// Get the resource key setting from the supplied <see cref="PipelineOptions"/> 
+        /// </summary>
+        /// <param name="options"></param>
+        /// <returns></returns>
+        public static string GetResourceKey(this PipelineOptions options)
+        {
+            var cloudConfig = options.GetElementConfig(nameof(CloudRequestEngine));
+            cloudConfig.BuildParameters.TryGetValue(RESOURCE_KEY_SETTING_NAME,
+                out var resourceKeyObj);
+            return resourceKeyObj?.ToString();
+        }
+
+        /// <summary>
+        /// Set the resource key setting in the supplied <see cref="PipelineOptions"/> instance
+        /// to the given value.
+        /// </summary>
+        /// <param name="options"></param>
+        /// <param name="resourceKey"></param>
+        public static void SetResourceKey(this PipelineOptions options, string resourceKey)
+        {
+            var cloudConfig = options.GetElementConfig(nameof(IpiCloudEngine));
+            cloudConfig.BuildParameters[RESOURCE_KEY_SETTING_NAME] = resourceKey;
+        }
+
+        /// <summary>
+        /// Set the cloud end point setting in the supplied <see cref="PipelineOptions"/> instance
+        /// to the given value.
+        /// </summary>
+        /// <param name="options"></param>
+        /// <param name="cloudEndPoint"></param>
+        public static void SetCloudEndPoint(this PipelineOptions options, string cloudEndPoint)
+        {
+            var cloudConfig = options.GetElementConfig(nameof(IpiCloudEngine));
+            cloudConfig.BuildParameters[CLOUD_END_POINT_SETTING_NAME] = cloudEndPoint;
+        }
+    }
+}
