+++ Examples/FiftyOne.IpIntelligence.Examples.OnPremise/ExampleUtils.cs
@@ -1,0 +1,97 @@
+/* *********************************************************************
+ * This Original Work is copyright of 51 Degrees Mobile Experts Limited.
+ * Copyright 2025 51 Degrees Mobile Experts Limited, Davidson House,
+ * Forbury Square, Reading, Berkshire, United Kingdom RG1 3EU.
+ *
+ * This Original Work is licensed under the European Union Public Licence
+ * (EUPL) v.1.2 and is subject to its terms as set out below.
+ *
+ * If a copy of the EUPL was not distributed with this file, You can obtain
+ * one at https://opensource.org/licenses/EUPL-1.2.
+ *
+ * The 'Compatible Licences' set out in the Appendix to the EUPL (as may be
+ * amended by the European Commission) shall be deemed incompatible for
+ * the purposes of the Work and the provisions of the compatibility
+ * clause in Article 5 of the EUPL shall not apply.
+ *
+ * If using the Work as, or as part of, a network application, by
+ * including the attribution notice(s) required under Article 5 of the EUPL
+ * in the end user terms of the application under an appropriate heading,
+ * such notice(s) shall fulfill the requirements of that article.
+ * ********************************************************************* */
+
+using FiftyOne.IpIntelligence.Engine.OnPremise.FlowElements;
+using FiftyOne.Pipeline.Core.FlowElements;
+using Microsoft.Extensions.Logging;
+
+using static FiftyOne.IpIntelligence.Examples.ExampleUtils;
+
+namespace FiftyOne.IpIntelligence.Examples.OnPremise
+{
+    public static class ExampleUtils
+    {
+
+        /// <summary>
+        /// Get information about the specified data file
+        /// </summary>
+        /// <param name="dataFile"></param>
+        /// <param name="engineBuilder"></param>
+        public static DataFileInfo GetDataFileInfo(string dataFile,
+            IpiOnPremiseEngineBuilder engineBuilder)
+        {
+            DataFileInfo result = new DataFileInfo();
+
+            using (var engine = engineBuilder
+                .Build(dataFile, false))
+            {
+                result = GetDataFileInfo(engine);
+            }
+
+            return result;
+        }
+
+        /// <summary>
+        /// Get information about the data file used by the specified engine
+        /// </summary>
+        /// <param name="engine"></param>
+        public static DataFileInfo GetDataFileInfo(IpiOnPremiseEngine engine)
+        {
+            DataFileInfo result = new DataFileInfo();
+            result.PublishDate = engine.DataFiles[0].DataPublishedDateTime;
+            result.Tier = engine.DataSourceTier;
+            result.Filepath = engine.DataFiles[0].DataFilePath;
+            return result;
+        }
+
+        /// <summary>
+        /// Display information about the data file and log warnings if specific requirements
+        /// are not met.
+        /// </summary>
+        /// <param name="pipeline"></param>
+        /// <param name="logger"></param>
+        public static void CheckDataFile(IPipeline pipeline, ILogger logger)
+        {
+            // Get the 'engine' element within the pipeline that performs IP Intelligence.
+            // We can use this to get details about the data file as well as meta-data describing
+            // things such as the available properties.
+            var engine = pipeline.GetElement<IpiOnPremiseEngine>();
+            CheckDataFile(engine, logger);
+        }
+
+        /// <summary>
+        /// Display information about the data file and log warnings if specific requirements
+        /// are not met.
+        /// </summary>
+        /// <param name="engine"></param>
+        /// <param name="logger"></param>
+        public static void CheckDataFile(IpiOnPremiseEngine engine, ILogger logger)
+        {
+            if (engine != null)
+            {
+                var info = GetDataFileInfo(engine);
+                LogDataFileInfo(info, logger);
+                LogDataFileStandardWarnings(info, logger);
+            }
+        }
+    }
+}
