+++ Examples/FiftyOne.IpIntelligence.Examples/FiftyOne.IpIntelligence.Examples.csproj
@@ -7,7 +7,7 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="FiftyOne.IpIntelligence" Version="4.5.0-alpha.120" />
+    <PackageReference Include="FiftyOne.IpIntelligence.Shared" Version="4.5.0-alpha.120" />
     <PackageReference Include="FiftyOne.Pipeline.Core" Version="4.5.0-alpha.28" />
     <PackageReference Include="YamlDotNet" Version="16.3.0" />
   </ItemGroup>
