+++ Examples/OnPremise/Framework-Web/Framework-Web.csproj
@@ -113,12 +113,6 @@
     </None>
   </ItemGroup>
   <ItemGroup>
-    <ProjectReference Include="..\..\FiftyOne.IpIntelligence.Examples\FiftyOne.IpIntelligence.Examples.csproj">
-      <Project>{6e062311-b25d-42f3-978f-6868ced67942}</Project>
-      <Name>FiftyOne.IpIntelligence.Examples</Name>
-    </ProjectReference>
-  </ItemGroup>
-  <ItemGroup>
     <PackageReference Include="FiftyOne.Pipeline.JavaScriptBuilder" Version="4.5.0-alpha.28" />
     <PackageReference Include="FiftyOne.Pipeline.Web" Version="4.5.0-alpha.28" />
     <PackageReference Include="Microsoft.Extensions.Configuration" Version="6.0.2" />
@@ -131,6 +125,16 @@
     <PackageReference Include="Newtonsoft.Json.Schema" Version="3.0.16" />
     <PackageReference Include="System.Memory" Version="4.6.3" />
     <PackageReference Include="System.Text.Json" Version="9.0.8" />
+  </ItemGroup>
+  <ItemGroup>
+    <ProjectReference Include="..\..\FiftyOne.IpIntelligence.Examples.OnPremise\FiftyOne.IpIntelligence.Examples.OnPremise.csproj">
+      <Project>{ed6b5c33-946e-2d4c-a7c2-0f606f1f9887}</Project>
+      <Name>FiftyOne.IpIntelligence.Examples.OnPremise</Name>
+    </ProjectReference>
+    <ProjectReference Include="..\..\FiftyOne.IpIntelligence.Examples\FiftyOne.IpIntelligence.Examples.csproj">
+      <Project>{6e062311-b25d-42f3-978f-6868ced67942}</Project>
+      <Name>FiftyOne.IpIntelligence.Examples</Name>
+    </ProjectReference>
   </ItemGroup>
   <PropertyGroup>
     <VisualStudioVersion Condition="'$(VisualStudioVersion)' == ''">10.0</VisualStudioVersion>
