+++ Examples/OnPremise/GettingStarted-Web/GettingStarted-Web.csproj
@@ -20,13 +20,14 @@
   </ItemGroup>
 
   <ItemGroup>
+	<PackageReference Include="FiftyOne.IpIntelligence" Version="4.5.0-alpha.120" />
 	<PackageReference Include="FiftyOne.Pipeline.Web" Version="4.5.0-alpha.28" />
 	<PackageReference Include="Microsoft.AspNetCore.Mvc.Razor.RuntimeCompilation" Version="8.0.19" />
 	<PackageReference Include="System.Text.Json" Version="8.0.6" />
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\..\FiftyOne.IpIntelligence.Examples\FiftyOne.IpIntelligence.Examples.csproj" />
+    <ProjectReference Include="..\..\FiftyOne.IpIntelligence.Examples.OnPremise\FiftyOne.IpIntelligence.Examples.OnPremise.csproj" />
   </ItemGroup>
 
 </Project>
