+++ FiftyOne.IpIntelligence.Examples.sln
@@ -41,6 +41,10 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "GettingStarted-Web", "Examp
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Metadata-Console", "Examples\Cloud\Metadata-Console\Metadata-Console.csproj", "{C89D3693-76BF-F318-AF28-E4206912FD74}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "FiftyOne.IpIntelligence.Examples.OnPremise", "Examples\FiftyOne.IpIntelligence.Examples.OnPremise\FiftyOne.IpIntelligence.Examples.OnPremise.csproj", "{ED6B5C33-946E-2D4C-A7C2-0F606F1F9887}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "FiftyOne.IpIntelligence.Examples.Cloud", "Examples\FiftyOne.IpIntelligence.Examples.Cloud\FiftyOne.IpIntelligence.Examples.Cloud.csproj", "{92BAF1E2-9A11-1EB1-0983-21DC7D2938BD}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		CoreDebug|Any CPU = CoreDebug|Any CPU
@@ -417,6 +421,54 @@ Global
 		{C89D3693-76BF-F318-AF28-E4206912FD74}.Release|ARM64.Build.0 = Release|ARM64
 		{C89D3693-76BF-F318-AF28-E4206912FD74}.Release|x64.ActiveCfg = Release|x64
 		{C89D3693-76BF-F318-AF28-E4206912FD74}.Release|x64.Build.0 = Release|x64
+		{ED6B5C33-946E-2D4C-A7C2-0F606F1F9887}.CoreDebug|Any CPU.ActiveCfg = Debug|Any CPU
+		{ED6B5C33-946E-2D4C-A7C2-0F606F1F9887}.CoreDebug|Any CPU.Build.0 = Debug|Any CPU
+		{ED6B5C33-946E-2D4C-A7C2-0F606F1F9887}.CoreDebug|ARM64.ActiveCfg = Debug|ARM64
+		{ED6B5C33-946E-2D4C-A7C2-0F606F1F9887}.CoreDebug|ARM64.Build.0 = Debug|ARM64
+		{ED6B5C33-946E-2D4C-A7C2-0F606F1F9887}.CoreDebug|x64.ActiveCfg = Debug|x64
+		{ED6B5C33-946E-2D4C-A7C2-0F606F1F9887}.CoreDebug|x64.Build.0 = Debug|x64
+		{ED6B5C33-946E-2D4C-A7C2-0F606F1F9887}.CoreRelease|Any CPU.ActiveCfg = Release|Any CPU
+		{ED6B5C33-946E-2D4C-A7C2-0F606F1F9887}.CoreRelease|Any CPU.Build.0 = Release|Any CPU
+		{ED6B5C33-946E-2D4C-A7C2-0F606F1F9887}.CoreRelease|ARM64.ActiveCfg = Release|ARM64
+		{ED6B5C33-946E-2D4C-A7C2-0F606F1F9887}.CoreRelease|ARM64.Build.0 = Release|ARM64
+		{ED6B5C33-946E-2D4C-A7C2-0F606F1F9887}.CoreRelease|x64.ActiveCfg = Release|x64
+		{ED6B5C33-946E-2D4C-A7C2-0F606F1F9887}.CoreRelease|x64.Build.0 = Release|x64
+		{ED6B5C33-946E-2D4C-A7C2-0F606F1F9887}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{ED6B5C33-946E-2D4C-A7C2-0F606F1F9887}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{ED6B5C33-946E-2D4C-A7C2-0F606F1F9887}.Debug|ARM64.ActiveCfg = Debug|ARM64
+		{ED6B5C33-946E-2D4C-A7C2-0F606F1F9887}.Debug|ARM64.Build.0 = Debug|ARM64
+		{ED6B5C33-946E-2D4C-A7C2-0F606F1F9887}.Debug|x64.ActiveCfg = Debug|x64
+		{ED6B5C33-946E-2D4C-A7C2-0F606F1F9887}.Debug|x64.Build.0 = Debug|x64
+		{ED6B5C33-946E-2D4C-A7C2-0F606F1F9887}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{ED6B5C33-946E-2D4C-A7C2-0F606F1F9887}.Release|Any CPU.Build.0 = Release|Any CPU
+		{ED6B5C33-946E-2D4C-A7C2-0F606F1F9887}.Release|ARM64.ActiveCfg = Release|ARM64
+		{ED6B5C33-946E-2D4C-A7C2-0F606F1F9887}.Release|ARM64.Build.0 = Release|ARM64
+		{ED6B5C33-946E-2D4C-A7C2-0F606F1F9887}.Release|x64.ActiveCfg = Release|x64
+		{ED6B5C33-946E-2D4C-A7C2-0F606F1F9887}.Release|x64.Build.0 = Release|x64
+		{92BAF1E2-9A11-1EB1-0983-21DC7D2938BD}.CoreDebug|Any CPU.ActiveCfg = Debug|Any CPU
+		{92BAF1E2-9A11-1EB1-0983-21DC7D2938BD}.CoreDebug|Any CPU.Build.0 = Debug|Any CPU
+		{92BAF1E2-9A11-1EB1-0983-21DC7D2938BD}.CoreDebug|ARM64.ActiveCfg = Debug|ARM64
+		{92BAF1E2-9A11-1EB1-0983-21DC7D2938BD}.CoreDebug|ARM64.Build.0 = Debug|ARM64
+		{92BAF1E2-9A11-1EB1-0983-21DC7D2938BD}.CoreDebug|x64.ActiveCfg = Debug|x64
+		{92BAF1E2-9A11-1EB1-0983-21DC7D2938BD}.CoreDebug|x64.Build.0 = Debug|x64
+		{92BAF1E2-9A11-1EB1-0983-21DC7D2938BD}.CoreRelease|Any CPU.ActiveCfg = Release|Any CPU
+		{92BAF1E2-9A11-1EB1-0983-21DC7D2938BD}.CoreRelease|Any CPU.Build.0 = Release|Any CPU
+		{92BAF1E2-9A11-1EB1-0983-21DC7D2938BD}.CoreRelease|ARM64.ActiveCfg = Release|ARM64
+		{92BAF1E2-9A11-1EB1-0983-21DC7D2938BD}.CoreRelease|ARM64.Build.0 = Release|ARM64
+		{92BAF1E2-9A11-1EB1-0983-21DC7D2938BD}.CoreRelease|x64.ActiveCfg = Release|x64
+		{92BAF1E2-9A11-1EB1-0983-21DC7D2938BD}.CoreRelease|x64.Build.0 = Release|x64
+		{92BAF1E2-9A11-1EB1-0983-21DC7D2938BD}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{92BAF1E2-9A11-1EB1-0983-21DC7D2938BD}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{92BAF1E2-9A11-1EB1-0983-21DC7D2938BD}.Debug|ARM64.ActiveCfg = Debug|ARM64
+		{92BAF1E2-9A11-1EB1-0983-21DC7D2938BD}.Debug|ARM64.Build.0 = Debug|ARM64
+		{92BAF1E2-9A11-1EB1-0983-21DC7D2938BD}.Debug|x64.ActiveCfg = Debug|x64
+		{92BAF1E2-9A11-1EB1-0983-21DC7D2938BD}.Debug|x64.Build.0 = Debug|x64
+		{92BAF1E2-9A11-1EB1-0983-21DC7D2938BD}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{92BAF1E2-9A11-1EB1-0983-21DC7D2938BD}.Release|Any CPU.Build.0 = Release|Any CPU
+		{92BAF1E2-9A11-1EB1-0983-21DC7D2938BD}.Release|ARM64.ActiveCfg = Release|ARM64
+		{92BAF1E2-9A11-1EB1-0983-21DC7D2938BD}.Release|ARM64.Build.0 = Release|ARM64
+		{92BAF1E2-9A11-1EB1-0983-21DC7D2938BD}.Release|x64.ActiveCfg = Release|x64
+		{92BAF1E2-9A11-1EB1-0983-21DC7D2938BD}.Release|x64.Build.0 = Release|x64
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -439,6 +491,8 @@ Global
 		{3156366A-FBFF-15F5-81AB-07DC688804B2} = {CA60EA6E-1E38-462D-A4A9-02DD86E92784}
 		{DD0BEAAA-0A7B-5D73-21D5-07DA6E231EEA} = {CA60EA6E-1E38-462D-A4A9-02DD86E92784}
 		{C89D3693-76BF-F318-AF28-E4206912FD74} = {CA60EA6E-1E38-462D-A4A9-02DD86E92784}
+		{ED6B5C33-946E-2D4C-A7C2-0F606F1F9887} = {1D4011D4-6726-43F6-83B1-5AF22757F1CB}
+		{92BAF1E2-9A11-1EB1-0983-21DC7D2938BD} = {1D4011D4-6726-43F6-83B1-5AF22757F1CB}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {7346AAA8-ADE3-458F-8EDB-925B1505A7A8}
